+++ Incomes.h
@@ -11,9 +11,9 @@ class Incomes
     int userId;
     int date;
     string item;
-    float amount;
+    double amount;
 public:
-    Incomes(int incomeId=0, int userId=0, int date = 0, string item = "", float amount = 0)
+    Incomes(int incomeId=0, int userId=0, int date = 0, string item = "", double amount = 0)
     {
         this -> incomeId = incomeId;
         this -> userId = userId;
@@ -25,13 +25,13 @@ public:
     void setUserId(int newUserId);
     void setDate(int newDate);
     void setItem(string newItem);
-    void setAmount(float newAmount);
+    void setAmount(double newAmount);
 
     int getIncomeId();
     int getUserId();
     int getDate();
     string getItem();
-    float getAmount();
+    double getAmount();
 };
 
 #endif
